add delete error check

+++ restflaskmongo/mongo_new.py
@@ -82,6 +82,8 @@ class Star(Resource):
     def delete(self, name):
         stars = mongo.db.stars
         star = [star for star in stars.find() if star['name'] == name]
+        if len(star) == 0:
+            abort(404)
         stars.delete_many({ "name": name } )
         return {'remove': marshal(star, star_fields)}, 201
